refactor: improve api structure of pairs endpoints
- change pairs/by-address/{address} to pairs/{address}
- change pairs/swap-routes to /swap routes and move to individual controller
